backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
sse-starlette==2.2.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0

# OpenAI
openai==1.58.1

# Vector store
faiss-cpu==1.9.0.post1
numpy==2.2.1

# Scraping
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
tiktoken==0.8.0
